requirements.txt
# Core dependencies
Flask==3.0.0
python-dotenv==1.0.0

# Alert integrations
discord-webhook==1.3.1
python-telegram-bot==20.7
slack-webhook==1.0.7
tweepy==4.14.0

# Required by dependencies
requests==2.31.0
urllib3==2.0.7
certifi==2024.2.2
